client/requirements.txt
anvil_web3
eth_abi
PGPy
psutil
pycryptodome
pyyaml
web3
